Partially fixed an issue with Roll20 adding classes.
"Note that by default, any HTML classes defined in your HTML layout
which don't begin with "attr", "roll" or "repeating_" will be prefixed
with "sheet-". So for example, if you want to style some of your input
tags to be shorter than others and you define the HTML as <input
class='shortfield'>, when processing your layout it will be changed to
<input class='sheet-shortfield'>."

To make dealing with this easier, this commit adds code that activates
on the full page only that modifies the DOM in this specific way,
prepending "sheet-" all non-permited class names.