Remove broken css regex
This regex used to execute a processing step that Roll20 also executes
which converts ID CSS selectors to class based ones. However, the regex
was too crude, and I had to remove the processing step. Roll20 does it
automatically so the styles will just fail, or not work as expected.
Also update the README.